more correct redacted attribute matching
